refactor(terraform-module): add zod schemas and validate service discovery
