+++ lib/modules/datasource/terraform-module/base.ts
@@ -3,7 +3,8 @@ import { withCache } from '../../../util/cache/package/with-cache.ts';
 import type { HttpError } from '../../../util/http/index.ts';
 import { ensureTrailingSlash } from '../../../util/url.ts';
 import { Datasource } from '../datasource.ts';
-import type { ServiceDiscoveryResult } from './types.ts';
+import type { ServiceDiscoveryResult } from './schema.ts';
+import { ServiceDiscoveryResponse } from './schema.ts';
 
 const terraformId = 'terraform';
 
@@ -11,14 +12,17 @@ const terraformId = 'terraform';
 export abstract class TerraformDatasource extends Datasource {
   static id = terraformId;
 
+  static readonly terraformRegistryUrl = 'https://registry.terraform.io';
+
   private async _getTerraformServiceDiscoveryResult(
     registryUrl: string,
   ): Promise<ServiceDiscoveryResult> {
     const discoveryURL = TerraformDatasource.getDiscoveryUrl(registryUrl);
-    const serviceDiscovery = (
-      await this.http.getJsonUnchecked<ServiceDiscoveryResult>(discoveryURL)
-    ).body;
-    return serviceDiscovery;
+    const { body: res } = await this.http.getJson(
+      discoveryURL,
+      ServiceDiscoveryResponse,
+    );
+    return res;
   }
 
   getTerraformServiceDiscoveryResult(
@@ -40,11 +44,9 @@ export abstract class TerraformDatasource extends Datasource {
 
   override handleHttpErrors(err: HttpError): void {
     const failureCodes = ['EAI_AGAIN'];
-    // istanbul ignore if
     if (failureCodes.includes(err.code)) {
       throw new ExternalHostError(err);
     }
-    // istanbul ignore if
     if (err.response?.statusCode === 503) {
       throw new ExternalHostError(err);
     }
