+++ lib/modules/datasource/terraform-module/schema.ts
@@ -1,0 +1,9 @@
+import { z } from 'zod/v3';
+
+export const ServiceDiscoveryResponse = z.object({
+  'modules.v1': z.string().optional(),
+  'providers.v1': z.string().optional(),
+});
+
+export type ServiceDiscoveryResult = z.infer<typeof ServiceDiscoveryResponse>;
+export type ServiceDiscoveryEndpointType = keyof ServiceDiscoveryResult;
